requirements.txt
numpy==1.14.3
stripe==2.35.0
watson_developer_cloud==1.4.0
zipcodes==1.0.5
Flask_WTF==0.14.2
pyodbc==4.0.23
requests==2.22.0
Flask_Uploads==0.2.1
itsdangerous==0.24
passlib==1.7.1
typing==3.6.2
pandas==0.23.0
WTForms==2.2.1
xmltodict==0.12.0
Flask_Mail==0.9.1
googleads==21.0.0
PyYAML==5.1.2
